// src/entities/types.ts
export type QuotationStatus = "draft" | "pending" | "approved" | "rejected" | "sent_to_client" | "deal" | "cancelled";
// ...
export type JenisLayanan =
    | "anti_rayap_injeksi"
    | "anti_rayap_pipanisasi"
    | "anti_rayap_baiting"
    | "anti_rayap_soil"
    | "anti_rayap_fumigasi"
    | "pest_spraying"
    | "pest_fogging"
    | "pest_rodent"
    | "pest_baiting"
    | "pest_umum"
;

export type KondisiBangunan = "pasca_konstruksi" | "pra_konstruksi" | "renovasi" | null;

export type TipeKontrak = "U" | "K" | "PH";
export type KategoriSurat = "AR" | "PCO";
// ...
export interface NomorSuratLog {
    id: string;
    noSurat: string;
    kategori: KategoriSurat;
    tipe: TipeKontrak;
    tipeLabel: string;
    jenisLayanan: JenisLayanan;
    kepada: string;
    byUid: string;
    byName: string;
    dibuat: Date;
    status: QuotationStatus;
    quoId: string | null;
    companyId: string;
    isManual?: boolean;
    keteranganManual?: string;
}